+++ singularity-eos/eos/get_sg_eos.cpp
@@ -48,7 +48,9 @@ int get_sg_eos( // sizing information
     // per material quantities
     double *frac_mass, double *frac_vol, double *frac_ie,
     // optional per material quantities
-    double *frac_bmod, double *frac_dpde, double *frac_cv) {
+    double *frac_bmod, double *frac_dpde, double *frac_cv,
+    // Mass fraction cutoff for PTE
+    double mass_frac_cutoff) {
   // printBacktrace();
   // kernel return value will be the number of failures
   int ret{0};
@@ -145,15 +147,16 @@ int get_sg_eos( // sizing information
   // declare init and final functors
   auto input_int_enum = static_cast<input_condition>(input_int);
   init_functor i_func;
-  final_functor f_func(temp_v, press_v, sie_v, bmod_v, cv_v, dpde_v, pte_mats, press_pte,
-                       vfrac_pte, temp_pte, sie_pte, frac_mass_v, frac_ie_v, frac_vol_v,
-                       vol_v, eos_v, pte_idxs, rho_pte, frac_bmod_v, frac_cv_v,
-                       frac_dpde_v, nmat, do_frac_bmod, do_frac_cv, do_frac_dpde);
+  final_functor f_func(spvol_v, temp_v, press_v, sie_v, bmod_v, cv_v, dpde_v, pte_mats,
+                       press_pte, vfrac_pte, temp_pte, sie_pte, frac_mass_v, frac_ie_v,
+                       frac_vol_v, vol_v, eos_v, pte_idxs, rho_pte, frac_bmod_v,
+                       frac_cv_v, frac_dpde_v, nmat, do_frac_bmod, do_frac_cv,
+                       do_frac_dpde);
   // only initialize init functor when needed
   if (input_int_enum != input_condition::P_T_INPUT) {
     i_func = init_functor(frac_mass_v, pte_idxs, eos_offsets_v, frac_vol_v, frac_ie_v,
                           pte_mats, vfrac_pte, sie_pte, temp_pte, press_pte, rho_pte,
-                          spvol_v, temp_v, press_v, sie_v, nmat);
+                          spvol_v, temp_v, press_v, sie_v, nmat, mass_frac_cutoff);
   }
 
   // create helper lambdas to reduce code duplication