+++ singularity-eos/eos/get_sg_eos_p_t.cpp
@@ -1,5 +1,5 @@
 //------------------------------------------------------------------------------
-// © 2021-2023. Triad National Security, LLC. All rights reserved.  This
+// © 2021-2024. Triad National Security, LLC. All rights reserved.  This
 // program was produced under U.S. Government contract 89233218CNA000001
 // for Los Alamos National Laboratory (LANL), which is operated by Triad
 // National Security, LLC for the U.S.  Department of Energy/National
@@ -85,7 +85,7 @@ void get_sg_eos_p_t(const char *name, int ncell, int nmat, indirection_v &offset
           vfrac_pte(tid, m) /= vfrac_tot;
         }
         // assign remaining outputs
-        f_func(i, tid, nmat, mass_sum, 0.0, 0.0, 0.0, cache);
+        f_func(i, tid, nmat, mass_sum, 0.0, 0.0, 0.0, true, cache);
         // assign max pressure
         pmax_v(i) = press_v(i) > pmax_v(i) ? press_v(i) : pmax_v(i);
         // release the token used for scratch arrays